backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
gunicorn==20.1.0
firecrawl==1.0.0
pydantic==1.9.0
python-dotenv==0.19.2
